requirements.txt
numpy
pandas
matplotlib
seaborn
sphinx-rtd-theme
python-dotenv
psycopg2
SQLAlchemy
python-slugify
selenium
wordcloud
Pillow
squarify
requests
matplotlib